+++ Source/ColonyManagerRedux/Helpers/UI/StockpileGUI.cs
@@ -10,9 +10,12 @@ namespace ColonyManagerRedux;
 [HotSwappable]
 public static class StockpileGUI
 {
+    public const int StockPilesPerRow = 2;
+
     private static List<Texture2D>? textures;
 
-    public static void DoStockpileSelectors(Rect rect, ref Zone_Stockpile? stockpile, Map map)
+    private static Vector2 _scrollPosition;
+    public static float DoStockpileSelectors(Vector2 position, float width, ref Zone_Stockpile? activeStockpile, Map map)
     {
         if (map == null)
         {
@@ -23,43 +26,64 @@ public static class StockpileGUI
         var allStockpiles = map.zoneManager.AllZones.OfType<Zone_Stockpile>().ToList();
 
         // count + 1 for all stockpiles
-        var areaCount = allStockpiles.Count + 1;
+        var stockPileCount = allStockpiles.Count + 1;
+        int rowCount = (int)Math.Ceiling((double)stockPileCount / StockPilesPerRow);
+        bool needsScrollbars = rowCount > 3;
+        float viewWidth = needsScrollbars ? width - 16f : width;
+        var widthPerCell = viewWidth / StockPilesPerRow;
 
         // create colour swatch
-        if (textures == null || textures.Count != areaCount - 1)
+        if (textures == null || textures.Count != stockPileCount - 1)
         {
             CreateTextures(allStockpiles);
         }
 
         mouseOverZone = null;
 
-        var widthPerCell = rect.width / areaCount;
         Text.WordWrap = false;
         Text.Font = GameFont.Tiny;
-        var nullAreaRect = new Rect(rect.x, rect.y, widthPerCell, rect.height);
-        DoZoneSelector(nullAreaRect, ref stockpile, null, BaseContent.GreyTex);
-        var areaIndex = 1;
-        for (var j = 0; j < allStockpiles.Count; j++)
+
+        Widgets.BeginScrollView(
+            new Rect(position.x, position.y, width, 3 * Constants.ListEntryHeight),
+            ref _scrollPosition,
+            new Rect(position.x, position.y, viewWidth, rowCount * Constants.ListEntryHeight),
+            needsScrollbars);
+
+        for (var j = 0; j < stockPileCount; j++)
         {
-            var xOffset = areaIndex * widthPerCell;
-            var stockpileRect = new Rect(rect.x + xOffset, rect.y, widthPerCell, rect.height);
-            DoZoneSelector(stockpileRect, ref stockpile, allStockpiles[j], textures[j]);
-            areaIndex++;
+            if (j == 0)
+            {
+                var nullAreaRect = new Rect(position.x, position.y, widthPerCell, Constants.ListEntryHeight);
+                DoZoneSelector(nullAreaRect, ref activeStockpile, null, BaseContent.GreyTex);
+            }
+            else
+            {
+                var stockpileRect = new Rect(
+                    position.x + j % StockPilesPerRow * widthPerCell,
+                    position.y + j / StockPilesPerRow * Constants.ListEntryHeight,
+                    widthPerCell, Constants.ListEntryHeight);
+                DoZoneSelector(stockpileRect, ref activeStockpile, allStockpiles[j - 1], textures[j - 1]);
+            }
         }
+
+        Widgets.EndScrollView();
 
         Text.WordWrap = true;
         Text.Font = GameFont.Small;
+
+        return rowCount * Constants.ListEntryHeight;
     }
 
 
     private static bool addedPostDrawSelectionOverlaysAction;
     private static Zone? mouseOverZone;
 
+    private static readonly Color MouseOverColor = new(.75f, .75f, .75f);
     private static void PostDrawSelectionOverlays()
     {
         if (mouseOverZone != null && !Find.Selector.IsSelected(mouseOverZone))
         {
-            GenDraw.DrawFieldEdges(mouseOverZone.Cells, Color.gray);
+            GenDraw.DrawFieldEdges(mouseOverZone.Cells, MouseOverColor);
         }
     }
 
@@ -87,7 +111,7 @@ public static class StockpileGUI
     }
 
     // RimWorld.AreaAllowedGUI
-    private static void DoZoneSelector(Rect rect, ref Zone_Stockpile? zoneAllowed, Zone_Stockpile? zone,
+    private static void DoZoneSelector(Rect rect, ref Zone_Stockpile? activeStockpile, Zone_Stockpile? zone,
                                         Texture2D tex)
     {
         if (!addedPostDrawSelectionOverlaysAction)
@@ -101,10 +125,11 @@ public static class StockpileGUI
         Text.Anchor = TextAnchor.MiddleLeft;
         var label = zone?.label ?? "Any stockpile";
         var innerRect = rect;
-        innerRect.xMin += 3f;
+        innerRect.xMin += 4f;
+        innerRect.xMax -= 4f;
         innerRect.yMin += 2f;
-        Widgets.Label(innerRect, label);
-        if (zoneAllowed == zone)
+        Widgets.Label(innerRect, label.Truncate(innerRect.width));
+        if (activeStockpile == zone)
         {
             Widgets.DrawBox(rect, 2);
         }
@@ -124,9 +149,9 @@ public static class StockpileGUI
             }
 
             if (Input.GetMouseButton(0) &&
-                 zoneAllowed != zone)
+                 activeStockpile != zone)
             {
-                zoneAllowed = zone;
+                activeStockpile = zone;
                 SoundDefOf.Designate_DragStandard_Changed.PlayOneShotOnCamera();
             }
         }